requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
Pillow
python-decouple
djangorestframework-simplejwt
celery
redis
django-celery-beat
qrcode
reportlab
WeasyPrint
django-extensions
factory-boy
pytest-django
pytest
coverage
flake8
black
gunicorn
whitenoise
psycopg2-binary
django-environ
django-storages
boto3
cryptography
scikit-learn
pandas
numpy
matplotlib
seaborn